+++ abnormalityAccept/src/main/java/com/abnormality/abnormalityaccept/entity/User.java
@@ -9,13 +9,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class User {
 
-    private Integer id;
+    private Long id;
     private String username;
     private String password;
     private String email;
-    private Integer inviterId;
-    private Integer leaderId;
-    private Integer facilityId;
+    private Long inviterId;
+    private Long leaderId;
+    private Long facilityId;
     private String introduction ;
     private Integer level;
